backend/requirements.txt
python-dotenv=0.16.0
Flask==1.1.2
Flask-Caching==1.9.0
Flask-JWT-Extended==4.1.0
Flask-Bcrypt==0.7.1
numpy==1.19.5
boto3==1.17.39
requests==2.25.1
haversine==2.3.0